Overhaul how Wemo Connect discovers and registers devices
Wemo's apparently have two MAC addresses. This was causing problems in
the connect app because SSDP messages were coming in with on MAC but
setup.xml returns a different MAC. Now the only thing SSDP messages are
used for is to get a device IP, and all registration is handled using
data from setup.xml.